+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * [#580](https://github.com/mozilla/glean.js/pull/580): Limit size of pings database to 250 pings or 10MB.
 * [#580](https://github.com/mozilla/glean.js/pull/580): BUGFIX: Pending pings at startup up are uploaded from oldest to newest.
 * [#607](https://github.com/mozilla/glean.js/pull/607): Record an error when incoherent timestamps are calculated for events after a restart.
+* [#630](https://github.com/mozilla/glean.js/pull/630): Accept booleans and numbers as event extras.
 
 # v0.18.1 (2021-07-22)